hashtag/requirements.txt
python-datamuse
beautifulsoup4
djangorestframework
pillow
brainyquote